pyproject.toml
[tool.poetry]
name = "ai-news"
version = "0.1.0"
description = "Latest AI news with chat assistant"
authors = ["Victor I. Afolabi <javafolabi@gmail.com>"]
maintainers = [
  "Victor I. Afolabi <javafolabi@gmail.com>",
]

readme = "README.md"
license = "MIT"

repository = "https://github.com/victor-iyi/ai-news"
keywords = ["ai-news", "news", "assistant"]

classifiers = [
  "Intended Audience :: Developers",
  "Intended Audience :: Education",
  "Intended Audience :: Science/Research",
  "License :: OSI Approved :: MIT License",
  "Natural Language :: English",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: Implementation :: CPython",
  "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
packages = [{include = "ai_news", from = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
newsapi-python = "^0.2.7"


[tool.poetry.group.dev.dependencies]
docformatter = {version = "^1.7.5", extras = ["tomli"]}
pydocstyle = {version = "^6.3.0", extras = ["tomli"]}
mypy = "^1.10.0"
pre-commit = "^3.7.0"
flake8 = "^7.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


# Docformatter.
[tool.docformatter]
in-place = true
recursive = true
blank = true
close-quotes-on-newline = true


# Style hinting.
[tool.mypy]
files = ["src/", "examples/"]
exclude = ["tests/", "venv/", ".venv/"]
namespace_packages = true
explicit_package_bases = true
strict = true
check_untyped_defs = true
ignore_missing_imports = true
disallow_any_generics = true
disallow_incomplete_defs = true
implicit_reexport = false
no_implicit_optional = true
show_error_codes = true
show_column_numbers = true
# show_error_content = true
warn_unused_configs = true
warn_unused_ignores = true
# warn_redundant_cast = true
warn_return_any = true
warn_unreachable = true
disallow_subclassing_any = false
disallow_untyped_decorators = false


[tool.pydocstyle]
convention = "google"
add_ignore = [
  # Missing docstring in public module.
  "D100",
  # Missing docstring in public class.
  "D101",
  # Missing docstring in public package.
  "D104",
  # No blank line allowed after function docstring.
  "D202",
  # 1 blank line required between summary line and description.
  "D205",
  # First line should end with a period, question mark, or exclamation point.
  "D415",
]


[tool.ruff.format]
quote-style = "single"
